update exe, add unit test

+++ src/circle.cpp
@@ -10,7 +10,7 @@ void Circle::lineIntersect(const Line& line, vector<struct Position>& res_vector
 	long long delta_y2 = this->y - line.y1;
 	long long cross = delta_x1 * delta_y2 - delta_x2 * delta_y1;
 	long long norm = delta_x1 * delta_x1 + delta_y1 * delta_y1;
-	double denominator = sqrt(norm);
+	double denominator = sqrt((double)norm);
 	double distance = fabs(cross / denominator);
 	if (distance - this->r * 1.0 < eps) {
 		double r_ = (delta_x2 * delta_x1 + delta_y2 * delta_y1) * 1.0 / norm;
@@ -53,7 +53,7 @@ void Circle::circleIntersect(const Circle& circle, vector<struct Position>& res_
 	if (subs1 == 0 || sigma < 0) {
 		return;
 	} else {
-		double sigma1 = sqrt(sigma);
+		double sigma1 = sqrt((double)sigma);
 		long long subs2 = -this->r * this->r * this->x + this->r * this->r * circle.x + circle.r * circle.r * this->x - circle.r * circle.r * circle.x + this->x * this->x * this->x - this->x * this->x * circle.x - this->x * circle.x * circle.x + this->x * this->y * this->y - 2 * this->x * this->y * circle.y + this->x * circle.y * circle.y + circle.x * circle.x * circle.x + circle.x * this->y * this->y - 2 * circle.x * this->y * circle.y + circle.x * circle.y * circle.y;
 		long long subs3 = -this->r * this->r * this->y + this->r * this->r * circle.y + circle.r * circle.r * this->y - circle.r * circle.r * circle.y + this->x * this->x * this->y + this->x * this->x * circle.y - 2 * this->x * circle.x * this->y - 2 * this->x * circle.x * circle.y + circle.x * circle.x * this->y + circle.x * circle.x * circle.y + this->y * this->y * this->y - this->y * this->y * circle.y - this->y * circle.y * circle.y + circle.y * circle.y * circle.y;
 		res1.x = (subs2 - sigma1 * this->y + sigma1 * circle.y) / (2 * subs1);
